--- a/example/server.js
+++ b/example/server.js
@@ -5,17 +5,19 @@ const blueconfig = require('blueconfig')
 blueconfig.addFormat(require('blueconfig-format-with-validator').ipaddress)
 
 const conf = blueconfig({
-  ip: {
-    doc: 'The IP Address to bind.',
-    format: 'ipaddress',
-    default: '127.0.0.1',
-    env: 'IP_ADDRESS'
-  },
-  port: {
-    doc: 'The port to bind.',
-    format: 'int',
-    default: 0,
-    env: 'PORT'
+  http: {
+    ip: {
+      doc: 'The IP Address to bind.',
+      format: 'ipaddress',
+      default: '127.0.0.1',
+      env: 'IP_ADDRESS'
+    },
+    port: {
+      doc: 'The port to bind.',
+      format: 'int',
+      default: 0,
+      env: 'PORT'
+    }
   }
 }).merge(path.join(__dirname, 'config.json')).validate()
 
@@ -24,7 +26,7 @@ const server = http.createServer(function(req, res) {
   res.end('Hello World\n')
 })
 
-server.listen(conf.get('port'), conf.get('ip'), function() {
+server.listen(conf.get('http.port'), conf.get('http.ip'), function() {
   const address = server.address()
   console.log('running on http://%s:%d', address.address, address.port) // eslint-disable-line no-console
 })