--- a/packages/blueconfig/lib/performer/utils/validator.js
+++ b/packages/blueconfig/lib/performer/utils/validator.js
@@ -1,64 +1,15 @@
 const stringifyPath = require('objectpath').stringify
 
-const walk = require('./walk.js')
 const utils = require('./utils.js')
-const isObjNotNull = utils.isObjNotNull
 const unroot = utils.unroot
 
 const cvtError = require('./../../error.js')
+const PATH_INVALID = cvtError.PATH_INVALID
 const VALUE_INVALID = cvtError.VALUE_INVALID
 
 
-function flatten(obj, useProperties) {
-  const stack = Object.keys(obj).map((path) => [path])
-  const entries = []
-
-  while (stack.length) {
-    const path = stack.shift()
-    let node = walk(obj, path)
-    // Is an object not null and not an array
-    if (isObjNotNull(node) && !Array.isArray(node)) {
-      if (useProperties) {
-        if ('_cvtProperties' in node) {
-          node = node._cvtProperties
-          path.push('_cvtProperties')
-        } else {
-          entries.push([path, node])
-          continue
-        }
-      }
-      const children = Object.keys(node)
-
-      // Don't filter out empty objects
-      if (children.length > 0) {
-        children.forEach(function(child) {
-          stack.push(path.concat(child))
-        })
-        continue
-      }
-    }
-    entries.push([path, node])
-  }
-
-  const flattened = {}
-  entries.forEach(function(entry) {
-    let path = entry[0]
-    const val = entry[1]
-
-    if (Array.isArray(path) === false) throw new Error('errror : ' + path)
-
-    if (useProperties) {
-      path = path.filter((property) => property !== '_cvtProperties')
-    }
-
-    flattened[stringifyPath(path)] = val
-  })
-
-  return flattened
-}
-
 function validator(strictValidation) {
-  const instance = this._instance
+  const node = this._instance
   const schema = this._schema
 
   const errors = {
@@ -67,52 +18,51 @@ function validator(strictValidation) {
     missing: []
   }
 
-  const flatInstance = flatten(instance)
-  const flatSchema = flatten(schema._cvtProperties, true)
+  function applyValidation(node, schema, path) {
+    const nodeKeys = Object.keys(node || {})
+    const schemaKeys = Object.keys(schema)
 
-  Object.keys(flatSchema).forEach(function(name) {
-    const schemaItem = flatSchema[name]
-    let instanceItem = flatInstance[name]
-    if (!(name in flatInstance)) {
-      try {
-        instanceItem = walk(instance, name)
-      } catch (err) {
-        let message = 'config parameter "' + unroot(name) + '" missing from config, did you override its parent?'
-        if (err.lastPosition && err.type === 'PATH_INVALID') {
-          message += ` Because ${err.why}.`
+    const keys = [...nodeKeys, ...schemaKeys]
+
+    for (let i = 0, len = keys.length; i < len; i++) {
+      const name = keys[i]
+
+      if (keys.indexOf(name) === i) {
+        const fullpath = [...path, name]
+
+        if (schemaKeys.indexOf(name) === -1) {
+          // If schema[name] doesn't exist:
+          if (strictValidation) {
+            const err = new VALUE_INVALID("configuration param '" + unroot(stringifyPath(fullpath)) + "' not declared in the schema")
+            errors.undeclared.push(err)
+          }
+        } else if (nodeKeys.indexOf(name) !== -1) {
+          // If node[name] and schema[name] exist:
+          if (!schema[name]._cvtProperties) {
+            // Is a property:
+            if (schema[name].attributes.required || !(typeof schema[name].attributes.default === 'undefined' &&
+                  node[name] === schema[name].attributes.default)) {
+              try {
+                schema[name].validate(node[name])
+              } catch (err) {
+                errors.invalid_type.push(err)
+              }
+            }
+          } else {
+            applyValidation(node[name], schema[name]._cvtProperties, fullpath)
+          }
+        } else {
+          // If node[name] doesn't exist and schema[name] exists:
+          const fullname = unroot(stringifyPath(fullpath))
+          const notfound = new PATH_INVALID(fullname, path, name, node)
+          const error = new VALUE_INVALID(`config parameter "${fullname}" missing from config, did you override its parent? Because ${notfound.why}.`)
+          errors.missing.push(error)
         }
-        errors.missing.push(new VALUE_INVALID(message))
-        return
       }
     }
-    delete flatInstance[name]
-
-    // ignore nested keys of schema 'object' properties
-    if (schemaItem.attributes.format === 'object' || typeof schemaItem.attributes.default === 'object') {
-      Object.keys(flatInstance)
-        .filter(function(key) {
-          return key.lastIndexOf(name + '.', 0) === 0
-        }).forEach(function(key) {
-          delete flatInstance[key]
-        })
-    }
-
-    if (schemaItem.attributes.required || !(typeof schemaItem.attributes.default === 'undefined' &&
-          instanceItem === schemaItem.attributes.default)) {
-      try {
-        schemaItem.validate(instanceItem)
-      } catch (err) {
-        errors.invalid_type.push(err)
-      }
-    }
-  })
-
-  if (strictValidation) {
-    Object.keys(flatInstance).forEach(function(name) {
-      const err = new VALUE_INVALID("configuration param '" + unroot(name) + "' not declared in the schema")
-      errors.undeclared.push(err)
-    })
   }
+
+  applyValidation(node, schema._cvtProperties, [])
 
   return errors
 }