## Full comparaison

```
convict@5.2.0 x 10,971 ops/sec ±4.76% (81 runs sampled)          
blueconfig@latest x 17,373 ops/sec ±1.52% (86 runs sampled)
blueconfig@6.0.2 x 12,765 ops/sec ±1.10% (84 runs sampled) 
Fastest is blueconfig@latest                               
```

For: 
```js
const conf = blueconfig602({
  http: {
    ip: {
      doc: 'The IP Address to bind.',
      format: 'ipaddress',
      default: '127.0.0.1',
      env: 'IP_ADDRESS'
    },
    port: {
      doc: 'The port to bind.',
      format: 'int',
      default: 0,
      env: 'PORT'
    },
    test: {
      test: {
        demo: 'true'
      }
    }
  }
}).merge({http:{port:80}}).validate()
```

## Before validate :

```
convict@5.2.0 x 16,751 ops/sec ±5.04% (82 runs sampled)
blueconfig@latest x 22,855 ops/sec ±1.31% (85 runs sampled)
blueconfig@6.0.2 x 22,801 ops/sec ±1.54% (88 runs sampled)
Fastest is blueconfig@latest,blueconfig@6.0.2
```

## Validate :

```
convict@5.2.0 x 40,531 ops/sec ±1.59% (86 runs sampled)
blueconfig@latest x 134,203 ops/sec ±1.52% (90 runs sampled)
blueconfig@6.0.2 x 37,208 ops/sec ±1.18% (92 runs sampled)
Fastest is blueconfig@latest
```

<details>
  <summary>Spoiler warning</summary>
  
  ```javascript
const Benchmark = require('benchmark')

const suite = new Benchmark.Suite;

// add tests
const convict = require('convict')
const blueconfig = require('../node-blueconfig/packages/blueconfig/')
const blueconfig602 = require('blueconfig')

blueconfig.addFormat(require('blueconfig-format-with-validator').ipaddress)
blueconfig602.addFormat(require('blueconfig-format-with-validator').ipaddress)

const conf_convict = convict({
  http: {
    ip: {
      doc: 'The IP Address to bind.',
      format: 'ipaddress',
      default: '127.0.0.1',
      env: 'IP_ADDRESS'
    },
    port: {
      doc: 'The port to bind.',
      format: 'int',
      default: 0,
      env: 'PORT'
    },
    test: {
      test: {
        demo: 'true'
      }
    }
  }
}).load({http:{port:80}})


const conf_blueconfig = blueconfig({
  http: {
    ip: {
      doc: 'The IP Address to bind.',
      format: 'ipaddress',
      default: '127.0.0.1',
      env: 'IP_ADDRESS'
    },
    port: {
      doc: 'The port to bind.',
      format: 'int',
      default: 0,
      env: 'PORT'
    },
    test: {
      test: {
        demo: 'true'
      }
    }
  }
}).merge({http:{port:80}})


const conf_blueconfig602 = blueconfig602({
  http: {
    ip: {
      doc: 'The IP Address to bind.',
      format: 'ipaddress',
      default: '127.0.0.1',
      env: 'IP_ADDRESS'
    },
    port: {
      doc: 'The port to bind.',
      format: 'int',
      default: 0,
      env: 'PORT'
    },
    test: {
      test: {
        demo: 'true'
      }
    }
  }
}).merge({http:{port:80}})


suite
  .add('convict@5.2.0', function() {
    conf_convict.validate()
  })
  .add('blueconfig@latest', function() {
    conf_blueconfig.validate()
  })
  .add('blueconfig@6.0.2', function() {
    conf_blueconfig602.validate()
  })
  // add listeners
  .on('cycle', function(event) {
    // check value?
    console.log(String(event.target));
  })
  .on('error', function(event) {
    console.log(event.target.error.message);
  })
  .on('complete', function() {
    console.log('Fastest is ' + this.filter('fastest').map('name'));
  })
  // run async
  .run({
    'async': true
  });
  ```
  
</details>